changes location of form components in project dir, moves to components dir, adjusts imports/exports and ensures necessary styles are in place
