+++ src/pages/BootcampSearch/index.js
@@ -1,0 +1,14 @@
+import SearchForm from '../../components/SearchForm';
+import styles from './styles';
+
+const BootcampSearch = () => {
+  return (
+    <>
+      <h1>Search</h1>
+      <SearchForm />
+      {/* Search Results */}
+    </>
+  );
+};
+
+export default BootcampSearch;
