+++ src/pages/SignUp/index.js
@@ -1,8 +1,8 @@
 import { Container } from '@mui/material';
-import SignUpForm from './form';
+import { useIsDesktop } from '../../_helpers/breakpoint';
+import SignUpForm from '../../components/SignUpForm';
 import styles from './styles';
 import './styles.css';
-import { useIsDesktop } from '../../_helpers/breakpoint';
 
 const imageStyles = {
   width: '675px',
